// LibrarySystem_FrontEnd/src/app/models/book-info-manager.model.ts
export class bookInfoManagerModel{
    isbn: string;
    useNum: number;
    wareCount: number;
    searchNum: string;
    publishPlace: string;
    picture: string;
    origin: string;
    remark: string;
    owner: string;
    bookName: string;
    totalpage: number;
    publishtime:string;
    updatetime: Date;
    typeName: string="";
    summary: string;
    author: string;
    bid: number;
    primCost: number;
    rank: number;
    sellPrice: number;
    language: string;
    id: string;
    content: string;
    operator:string;
    name:string;
    location:string

    }

export class BookInfoAddModel{
    isbn: string;
    useNum: number;
    searchNum: string;
    publishPlace: string;
    picture: string;
    origin: string;
    remark: string;
    owner: string;
    bookName: string;
    totalpage: number;
    publishtime: string;
    updatetime: Date;
    typeName: string;
    summary: string;
    author: string;
    bid: number;
    primCost: number;
    rank: number;
    sellPrice: number;
    language: string;
    id: string;
    content: string;
    operator:string;
    // "collectionId": "string",
    }